[addons] move addon headers for instances to another place
This moves the headers to <kodi/addon-instance/..> to prevent a to big
amount of not needed folders after all instance types are changed.
