+++ xbmc/addons/kodi-addon-dev-kit/include/kodi/addon-instance/Screensaver.h
